+++ src/magnet2torrent/__main__.py
@@ -4,6 +4,8 @@ import ipaddress
 import logging
 import os
 
+from pathlib import Path
+
 from aiohttp import web
 
 from . import settings
@@ -20,7 +22,10 @@ def main():
         "--use-dht", help="Enable DHT", action="store_true", dest="use_dht"
     )
     parser.add_argument(
-        "--dht-state-file", help="Where to save DHT info", dest="dht_state_file"
+        "--dht-state-file", help="Where to save DHT info", dest="dht_state_file", type=str,
+    )
+    parser.add_argument(
+        "--torrent-cache-folder", help="Folder to cache torrent metadata into", dest="torrent_cache_folder", type=str,
     )
     subparsers = parser.add_subparsers(help="sub-command help", dest="command")
 
@@ -51,6 +56,16 @@ def main():
     fetch_subparser.add_argument("magnet", help="Magnet link")
 
     args = parser.parse_args()
+
+    if args.torrent_cache_folder:
+        torrent_cache_folder = Path(args.torrent_cache_folder)
+
+        if not torrent_cache_folder.exists():
+            os.makedirs(args.torrent_cache_folder)
+
+        if not torrent_cache_folder.is_dir():
+            print(f"Path {args.torrent_cache_folder} exists but is not a folder")
+            quit(1)
 
     if args.debug:
         logging.basicConfig(
@@ -87,12 +102,13 @@ def main():
 
         settings.SERVE_APIKEY = args.apikey
         settings.DHT_SERVER = dht_server
+        settings.TORRENT_CACHE_FOLDER = args.torrent_cache_folder
         app = web.Application()
         app.add_routes(routes)
         web.run_app(app, host=str(args.ip), port=args.port)
     elif args.command == "fetch":
         loop = asyncio.get_event_loop()
-        m2t = Magnet2Torrent(args.magnet, dht_server=dht_server)
+        m2t = Magnet2Torrent(args.magnet, dht_server=dht_server, torrent_cache_folder=args.torrent_cache_folder)
         try:
             filename, torrent_data = loop.run_until_complete(m2t.retrieve_torrent())
         except FailedToFetchException: