+++ src/magnet2torrent/exceptions.py
@@ -1,8 +1,2 @@
 class FailedToFetchException(Exception):
     """Unable to fetch torrent exception"""
-
-
-class MalformedMessage(Exception):
-    """
-    Message does not contain what is expected.
-    """