+++ src/magnet2torrent/__init__.py
@@ -1,4 +1,4 @@
 from .exceptions import FailedToFetchException
 from .magnet2torrent import Magnet2Torrent
 
-__version__ = "1.0.4"
+__version__ = "1.1.0"